feat: secondary button variant
